@@ -14,12 +14,14 @@
   "prefer-stable": true,
   "minimum-stability": "dev",
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1 <8.4",
     "ext-json": "*",
+    "ext-simplexml": "*",
     "symfony/yaml": "> 4"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.7.*|7.4.*|^9.5"
+    "phpunit/phpunit": "^9.6",
+    "vimeo/psalm": "^5.9"
   },
   "suggest": {
     "ext-curl": "*"
