@@ -2,21 +2,21 @@
 
 namespace ByJG\Serializer\Formatter;
 
-use ByJG\Serializer\SerializerObject;
+use ByJG\Serializer\Serialize;
 
 class JsonFormatter implements FormatterInterface
 {
 
     /**
-     * @param array|object $serializable
+     * @param object|array $serializable
      * @return string
      */
-    public function process($serializable)
+    public function process(object|array $serializable): string
     {
         if (is_array($serializable)) {
             return json_encode($serializable);
         }
 
-        return json_encode(SerializerObject::instance($serializable)->serialize());
+        return json_encode(Serialize::from($serializable)->toArray());
     }
 }
