@@ -1,7 +1,8 @@
 <?php
 
-namespace ByJG\Serializer;
+namespace Tests\Formatter;
 
+use ByJG\Serializer\Serialize;
 use PHPUnit\Framework\TestCase;
 use Tests\Sample\ModelGetter;
 use Tests\Sample\SampleModel;
@@ -41,7 +42,7 @@ class JsonFormatterTest extends TestCase
 
         $formatter = new JsonFormatter();
         $this->assertEquals('{"collection":[{"Id":10,"Name":"John"},{"Id":20,"Name":"Doe"}]}', $formatter->process($object));
-        $this->assertEquals('[{"Id":10,"Name":"John"},{"Id":20,"Name":"Doe"}]', $formatter->process(SerializerObject::instance($object->getCollection())->serialize()));
+        $this->assertEquals('[{"Id":10,"Name":"John"},{"Id":20,"Name":"Doe"}]', $formatter->process(Serialize::from($object->getCollection())->toArray()));
     }
 
 }
