@@ -1,7 +1,8 @@
 <?php
 
-namespace ByJG\Serializer;
+namespace Tests\Formatter;
 
+use ByJG\Serializer\Serialize;
 use PHPUnit\Framework\TestCase;
 use Tests\Sample\ModelGetter;
 use Tests\Sample\SampleModel;
@@ -58,7 +59,7 @@ class YamlFormatterTest extends TestCase
 
         $this->assertEquals(
             file_get_contents(__DIR__ . "/yaml4.yml"),
-            $formatter->process(SerializerObject::instance($object->getCollection())->serialize())
+            $formatter->process(Serialize::from($object->getCollection())->toArray())
         );
     }
 
