@@ -1,15 +1,19 @@
 <?php
 
-namespace ByJG\Serializer;
+namespace Tests\Serialize;
 
 use ByJG\Serializer\PropertyPattern\SnakeToCamelCase;
 use PHPUnit\Framework\TestCase;
+use stdClass;
 use Tests\Sample\ModelGetter;
+use Tests\Sample\ModelPropertyPattern;
+use Tests\Sample\ModelPropertyPatternAnnotation;
+use Tests\Sample\ModelPropertyPatternConstruct;
 use Tests\Sample\SampleModel;
 
 class BaseModelTest extends TestCase
 {
-    public function testBindFromObject()
+    public function testCopyFromObject()
     {
         $model = new ModelGetter(10, 'Testing');
 
@@ -19,10 +23,10 @@ class BaseModelTest extends TestCase
         $this->assertEquals("Testing", $object->getName());
     }
 
-    public function testBindFromStdClass()
+    public function testCopyFromStdClass()
     {
         // Matching exact property names
-        $model = new \stdClass();
+        $model = new stdClass();
         $model->Id = 10;
         $model->Name = "Testing";
 
@@ -32,7 +36,7 @@ class BaseModelTest extends TestCase
         $this->assertEquals("Testing", $object->getName());
 
         // Matching with different case letters
-        $model2 = new \stdClass();
+        $model2 = new stdClass();
         $model2->id = 10;
         $model2->name = "Testing";
 
@@ -42,7 +46,7 @@ class BaseModelTest extends TestCase
         $this->assertEquals("Testing", $object->getName());
     }
 
-    public function testBindFromArray()
+    public function testCopyFromArray()
     {
         $array = array(
             "Id" => 10,
@@ -55,41 +59,41 @@ class BaseModelTest extends TestCase
         $this->assertEquals("Testing", $object->getName());
     }
 
-   public function testPropertyPatternBind()
+   public function testPropertyPatternCopy()
    {
-       $obj = new \stdClass();
+       $obj = new stdClass();
        $obj->Id_Model = 10;
        $obj->Client_Name = 'Testing';
 
-       // Testing Without Property Bind
-       $object = new \Tests\Sample\ModelPropertyPattern();
-       $object->bindFrom($obj);
+       // Testing Without Property Copy
+       $object = new ModelPropertyPattern();
+       $object->copyFrom($obj);
 
        $this->assertEquals('', $object->getIdModel());
        $this->assertEquals('', $object->getClientName());
 
-       // Testing with Bind
-       $object = new \Tests\Sample\ModelPropertyPattern();
-       $object->bindFrom($obj, new SnakeToCamelCase());
+       // Testing with Copy
+       $object = new ModelPropertyPattern();
+       $object->copyFrom($obj, new SnakeToCamelCase());
 
        $this->assertEquals(10, $object->getIdModel());
        $this->assertEquals("Testing", $object->getClientName());
 
        // Testing Constructor
-       $object = new \Tests\Sample\ModelPropertyPattern($obj, new SnakeToCamelCase());
+       $object = new ModelPropertyPattern($obj, new SnakeToCamelCase());
 
        $this->assertEquals(10, $object->getIdModel());
        $this->assertEquals("Testing", $object->getClientName());
    }
 
-   public function testPropertyPatternBind_2()
+   public function testPropertyPatternCopy_2()
    {
        // Other Testing
-       $obj = new \stdClass();
+       $obj = new stdClass();
        $obj->IdModel = 10;
        $obj->ClientName = 'Testing';
 
-       $object = new \Tests\Sample\ModelPropertyPattern($obj);
+       $object = new ModelPropertyPattern($obj);
 
        $this->assertEquals(10, $object->getIdModel());
        $this->assertEquals("Testing", $object->getClientName());
@@ -98,7 +102,7 @@ class BaseModelTest extends TestCase
    /**
     * The current property pattern try do remove the underscore.
     */
-   public function testPropertyPatternBind_3()
+   public function testPropertyPatternCopy_3()
    {
        // Other Testing
        $obj = [
@@ -106,8 +110,8 @@ class BaseModelTest extends TestCase
            "birthdate" => "1974-01-26"
        ];
 
-       $object = new \Tests\Sample\ModelPropertyPattern();
-       $object->bindFrom($obj);
+       $object = new ModelPropertyPattern();
+       $object->copyFrom($obj);
 
        $this->assertEquals("1974-01-26", $object->getBirthdate());
        $this->assertEquals("Joao", $object->getClientName());
@@ -117,15 +121,15 @@ class BaseModelTest extends TestCase
     * The current property pattern try do remove the underscore.
     * The setPropertyPattern is done on constructor
     */
-   public function testPropertyPatternBind_4()
+   public function testPropertyPatternCopy_4()
    {
        // Other Testing
        $obj = [
            "birth_date" => "1974-01-26"
        ];
 
-       $object = new \Tests\Sample\ModelPropertyPatternConstruct();
-       $object->bindFrom($obj);
+       $object = new ModelPropertyPatternConstruct();
+       $object->copyFrom($obj);
 
        $this->assertEquals("1974-01-26", $object->getBirth_date());
    }
@@ -133,14 +137,14 @@ class BaseModelTest extends TestCase
    /**
     * The current property pattern try do remove the underscore.
     */
-   public function testPropertyPatternBind_5()
+   public function testPropertyPatternCopy_5()
    {
        // Other Testing
        $obj = [
            "birth_date" => "1974-01-26"
        ];
 
-       $object = new \Tests\Sample\ModelPropertyPatternAnnotation($obj);
+       $object = new ModelPropertyPatternAnnotation($obj);
 
        $this->assertEquals("1974-01-26", $object->getBirth_date());
    }
